Extract MongoJournalRepository and MongoEntryRepository
Second slice of splitting the monolithic Mongo repository: journals and
entries now live in their own per-aggregate repositories, extracted together
because they are entangled (entry reads derive their permissions from journal
access; journal deletion cascades into entries).

- MongoJournalRepository: journal data access shaped by the injected
  IReadScope, including the explicitly-unscoped GetJournalUnscoped primitive
  used by the write guards.
- MongoEntryRepository: entry data access. Takes no IReadScope - entry
  scoping is derived entirely from journal permissions via the (scoped)
  journal repository; SearchEntries/GetEntry remain documented unscoped
  primitives.
- FreeTextFilters: the shared free-text regex filter building, extracted from
  the base class for both repositories.
- MongoRepositoryBase is now a pure delegating facade holding only the
  composition and the virtual write-guard hooks; it dies entirely in the
  follow-up cleanup that turns the guards into decorators.

Behavior-preserving: public constructors, interfaces and DI wiring unchanged.
